requirements.txt
Appium-Python-Client==1.3.0
configparser